refactored code to fix main
