+++ Javascript-Node/public/lib/auth/firebase-config.js
