+++ Javascript-Node/public/lib/common/nav_control.js
