+++ Javascript-Node/public/lib/listings/render_listing.js
