+++ Javascript-Node/public/templates/create_listing.html
@@ -4,8 +4,8 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
   <title>Create Listing</title>
-  <link rel="stylesheet" href="../static/style.css" />
-  <link rel="stylesheet" href="../static/create_listing.css">
+  <link rel="stylesheet" href="../static/base.css" />
+  <link rel="stylesheet" href="../static/pages/create_listing.css">
     
   </style>
 </head>
@@ -52,7 +52,7 @@
     </form>
   </div>
 
-  <script type="module" src="../lib/create_listing.js"></script>
+  <script type="module" src="../lib/listings/create_listing.js"></script>
 
 </body>
 </html>
