+++ Javascript-Node/public/templates/dashboard.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
 
-    <link rel="stylesheet" href="../static/style.css" />
+    <link rel="stylesheet" href="../static/base.css" />
 </head>
 <body>
     <nav>
