+++ Javascript-Node/public/templates/index.html
@@ -4,9 +4,9 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>My JavaScript Web App</title>
-  <link rel="stylesheet" href="../static/style.css" />
-  <link rel="stylesheet" href="../static/listings.css" />
-  <script src="../lib/listings_script.js" defer></script>
+  <link rel="stylesheet" href="../static/base.css" />
+  <link rel="stylesheet" href="../static/pages/listings.css" />
+  <script src="../lib/listings/render_listing.js" defer></script>
 </head>
 <body>
   <nav>
@@ -24,7 +24,7 @@
 
   
   <!-- All cookie logic runs after DOM is ready -->
-  <script type="module" src="../lib/nav_control.js"></script>
+  <script type="module" src="../lib/common/nav_control.js"></script>
