+++ Javascript-Node/public/templates/login.html
@@ -5,8 +5,8 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>My JavaScript Web App</title>
-    <link rel="stylesheet" href="../static/style.css" />
-    <link rel="stylesheet" href="../static/register_login.css" />
+    <link rel="stylesheet" href="../static/base.css" />
+    <link rel="stylesheet" href="../static/pages/register_login.css" />
 </head>
 
 <body>
@@ -35,9 +35,8 @@
         </main>
     </div>
     <!-- Firebase setup -->
-    <script type="module" src="../lib/firebase-config.js"></script>
-    <script type="module" src="../lib/handle_login.js"></script>
-    <script type="module" src="../lib/login_dom.js"></script>
+    <script type="module" src="../lib/auth/firebase-config.js"></script>
+    <script type="module" src="../lib/login/handle_login.js"></script>
 </body>
 
 </html>
