+++ Javascript-Node/public/templates/sign_up.html
@@ -5,8 +5,8 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>My JavaScript Web App</title>
-  <link rel="stylesheet" href="../static/style.css" />
-  <link rel="stylesheet" href="../static/register_login.css" />
+  <link rel="stylesheet" href="../static/base.css" />
+  <link rel="stylesheet" href="../static/pages/register_login.css" />
 
 
 </head>
@@ -86,12 +86,9 @@
   <!-- JS Imports -->
   <!-- Importing CryptoJS using a CDN so that we can hash sensitive data -->
   <script src="https://cdnjs.cloudflare.com/ajax/libs/crypto-js/4.1.1/crypto-js.min.js"></script>
-  <!-- Firebase setup -->
-  <script type="module" src="../lib/firebase-config.js"></script>
-  <!-- Create a user -->
-  <script type="module" src="../lib/handle_signup.js"></script>
+
   <!-- DOM handling -->
-  <script type="module" src="../lib/signup_dom.js"></script>
+  <script type="module" src="../lib/sign_up/signup_dom.js"></script>
 
 </body>
